+++ Cheers/BarNames.txt
@@ -26,3 +26,12 @@ Thursday
 Friday
 3 pm - 6 pm
 END OF RECORD
+El Prez
+Pacific Beach
+Wednesday
+3 pm - 6 pm
+Thursday
+3 pm - 6 pm
+Friday
+3 pm - 6 pm
+END OF RECORD
