+++ Podfile
@@ -16,6 +16,8 @@ target 'Cheers' do
   # Date/Time Management
   pod 'SwiftDate', '~> 4.5.1'
   
+  # Progress HUD
+  pod 'SVProgressHUD'
 
   # Networking & Parsing
   pod 'SwiftyJSON'
